Fixed bug due to treating "time" as a program.
In many shells (including bash) "time" invokes a shell
command, not a separate program. The following simply
times the execution of a process that writes stderr to
a file blah:

$ time prog 2> blah

This is not what we want, which is the following:

$ ( time prog ) 2> blah

which writes the stderr output of the time command to blah.